refactor: move component related to comment, thread to src/components
