--- src/components/ArticleLoading.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react'
-import { Skeleton } from 'antd'
-
-interface Props {
-  loading: boolean
-}
-
-const ArticleLoading: React.FC<Props> = ({ loading }) => {
-  return <>{loading && <Skeleton active={true} />}</>
-}
-
-export default ArticleLoading
